renamed auth.js to hash.js
