+++ backend/hash.js
